Fix delete action and property read bugs (#191)
Fix NPE while performing a delete action on a non-existing workflow.

Fix syntax bug on property injection.

+++ workflow-bot-app/src/main/java/com/symphony/bdk/workflow/management/WorkflowAbstractAction.java
@@ -56,6 +56,9 @@ public abstract class WorkflowAbstractAction {
   }
 
   protected void deleteFile(String workflowId) {
+    if (!workflowExist(workflowId)) {
+      throw new NotFoundException(String.format("Workflow %s does not exist", workflowId));
+    }
     File file = deployer.workflowSwadlPath(workflowId).toFile();
     if (!file.delete()) {
       throw new NotFoundException(String.format("Workflow %s does not exist", workflowId));
